requirements.txt
playwright
beautifulsoup4
python-dotenv
geopy
nltk